requirements.txt
Pillow>=10.0.0
scikit-learn>=1.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
ultralytics>=8.3.0
PyMuPDF>=1.24.0
